mqtt-chat/requirements.txt
flask
flask-cors
mysql-connector-python
paho-mqtt